+++ js/form.js
@@ -1,6 +1,5 @@
 //Зависимость минимального значения и плейсхолдера поля «Цена за ночь» от типа жилья
-
-const MIN_COST_HOUSING = {
+const MinCostsHousing = {
   bungalow : 0,
   flat : 1000,
   house : 5000,
@@ -11,22 +10,8 @@ const priceHousingInput = document.querySelector('#price');
 const typeHousingInput = document.querySelector('#type');
 
 typeHousingInput.addEventListener('input', function() {
-  let value;
-
-  if (typeHousingInput.value === 'bungalow') {
-    value = MIN_COST_HOUSING.bungalow;
-  }
-  if (typeHousingInput.value === 'flat') {
-    value = MIN_COST_HOUSING.flat;
-  }
-  if (typeHousingInput.value === 'house') {
-    value = MIN_COST_HOUSING.house;
-  }
-  if (typeHousingInput.value === 'palace') {
-    value = MIN_COST_HOUSING.palace;
-  }
-  priceHousingInput.setAttribute('min', value);
-  priceHousingInput.setAttribute('placeholder', value);
+  priceHousingInput.setAttribute('min', MinCostsHousing[typeHousingInput.value]);
+  priceHousingInput.setAttribute('placeholder', MinCostsHousing[typeHousingInput.value]);
 });
 
 //Обработчики событий «Время заезда», «Время выезда»
