+++ js/main.js
@@ -1,3 +1,3 @@
 import './data.js';
 import './form.js';
-import './generationOfMarkupAd.js';
+import './generationofmarkupad.js';
